+++ tests/run.sh
@@ -8,6 +8,7 @@ ROOT_DIR="$(
 
 for test in \
     test_static.py \
+    test_uninstall.py \
     test_install_web.py \
     test_web_api.py \
     test_web_http.py \